results fix

+++ cor.py
@@ -3,24 +3,34 @@ import os
 import sys
 import json
 import shutil
+import pandas as pd
+
+from tqdm import tqdm
 
 dicts_path = 'files'
 cisbp_human_dict_path = os.path.join('source_files', 'TF_Information_human.txt')
 cisbp_mouse_dict_path = os.path.join('source_files', 'TF_Information_mouse.txt')
-info_dict_path = os.path.join('files', 'new_info.json')
+info_dict_path = os.path.join('files', 'filtered_info.json')
+initial_info_dict_path = os.path.join('files', 'info.json')
 result_path = 'ape_result'
 motif_dir = 'cisbp_pwms'
 
 dict_types = ['hocomoco', 'direct', 'inferred', 'tf_class_family', 'tf_class_subfamily']
-
 
 ape_path = os.path.expanduser('~/ape.jar')
 hocomoco_path = 'hocomoco_pwms'
 allowed_tfs = ['ANDR_MOUSE', 'CTCF_MOUSE']
 
 
+def read_xlsx_master():
+    return pd.read_excel(os.path.join('source_files', 'hocomoco_2021.xlsx'),
+                         engine='openpyxl')
+
+
 def check_dir_for_collection(tf, motif_collection, d_type):
     dir_path = os.path.join(result_path, tf + '_' + d_type)
+    if motif_collection is None:
+        return None
     if os.path.exists(dir_path):
         shutil.rmtree(dir_path)
     os.mkdir(dir_path)
@@ -52,25 +62,25 @@ def parse_output(exp, output):
     return pcm_name, best_motif
 
 
-def run_ape(exps, res_dir, jobs=10, additional_message=''):
+def run_ape(exps, res_dir, jobs=10):
     result = {}
-    exps_batches = [exps[i:i + jobs] for i in range(0, len(exps), jobs)]
+    if res_dir is not None:
+        exps_batches = [exps[i:i + jobs] for i in range(0, len(exps), jobs)]
 
-    for index, batch in enumerate(exps_batches, 1):
-        print('Doing {}/{} batch {}'.format(index, len(exps_batches), additional_message))
-        commands = [["java", '-cp', ape_path,
-                     'ru.autosome.macroape.ScanCollection',
-                     exp, res_dir, '--query-pcm', '-d', '1', '--all'] for exp in batch]
-        processes = [subprocess.Popen(command,
-                                      stdout=subprocess.PIPE,
-                                      stderr=subprocess.PIPE) for command in commands]
-        for process in processes:
-            process.wait()
-        for process, exp in zip(processes, batch):
-            res = process.stdout.read().decode('utf-8')
-            name, res = parse_output(exp, res)
-            result[name] = res
-    shutil.rmtree(res_dir)
+        for batch in exps_batches:
+            commands = [["java", '-cp', ape_path,
+                         'ru.autosome.macroape.ScanCollection',
+                         exp, res_dir, '--query-pcm', '-d', '1', '--all'] for exp in batch]
+            processes = [subprocess.Popen(command,
+                                          stdout=subprocess.PIPE,
+                                          stderr=subprocess.PIPE) for command in commands]
+            for process in processes:
+                process.wait()
+            for process, exp in zip(processes, batch):
+                res = process.stdout.read().decode('utf-8')
+                name, res = parse_output(exp, res)
+                result[name] = res
+        shutil.rmtree(res_dir)
     return result
 
 
@@ -83,46 +93,34 @@ def read_dicts():
 
 
 def transform_name(tf):
-    if tf == 'ANDR_HUMAN':
-        result = 'AR'
-    else:
-        result = tf[:-6] if tf.endswith('_HUMAN') else tf
-    return result
+    return tf
 
 
 def filter_pwms(motifs):
     words_tr = 50
     percent_tr = 0.25
     return [x for x in motifs if x['pcm_path']
-            and x['words'] >= words_tr and x['total'] and x['seqs']/x['total'] >= percent_tr]
+            and x['words'] >= words_tr and x['total'] and x['seqs'] / x['total'] >= percent_tr]
 
 
 def main(njobs=10):
     dicts = read_dicts()
     with open(info_dict_path) as info:
         info_dict = json.loads(info.readline())
-    hoco_dict = [x for x, y in dicts['hocomoco'].items() if len(y) > 0]
-    for index, tf_name in enumerate(hoco_dict):
-        tf = tf_name + '_HUMAN'
+    for tf in tqdm(info_dict.keys()):
         if os.path.exists(os.path.join('reports', tf + '.xlsx')):
             continue
         results = {}
-        if tf not in info_dict:
-            continue
         pwms = [x['pcm_path'] for x in filter_pwms(info_dict[tf])]
-        t_name = transform_name(tf)
         for d_type in dict_types:
-            if len(pwms) > 1000 and d_type != 'hocomoco':
-                continue
-            motif_collection = dicts[d_type].get(t_name)
-            if not motif_collection:
+            motif_collection = dicts[d_type].get(tf)
+            if not motif_collection and d_type != 'hocomoco':
                 continue
             res_dir = check_dir_for_collection(tf, motif_collection, d_type)
-            additional_message = 'for {} {} ({}/{})'.format(tf, d_type, index, len(hoco_dict))
-            ape_res = run_ape(pwms, res_dir, njobs, additional_message)
+            ape_res = run_ape(pwms, res_dir, njobs)
             results[d_type] = ape_res
         with open(os.path.join(result_path, tf + '.json'), 'w') as out:
-            json.dump(results, out)
+            json.dump(results, out, indent=2)
 
 
 if __name__ == '__main__':
